requirements.txt
django
djangorestframework
psycopg2
django-filter
python-dotenv
uritemplate
inflection
